Variance GWA Manhattan Plot

The function plots the variance GWA result for the given scan object.

# S3 method for vGWAS
plot(x, sig.threshold = NULL, low.log.p = 0,
pch = 16, cex = 0.6, col.manhattan = c("slateblue4", "olivedrab"),
col.sig.threshold = "darkgoldenrod", ...)

Arguments

x

a result object from vGWAS scan. It can be any list or data.frame that contains chromosome, marker.map, and p.value, with class = 'vGWAS'. See vGWAS.

sig.threshold

a numeric value giving the significance threshold for -log(pvalues, 10). If NULL, Bonferroni correction will be used.

low.log.p

a numeric value giving the lower limit of the -log(pvalues, 10) to plot.

pch

point character. See par.

cex

size of points. See par.

col.manhattan

two colors as a vector for the Manhattan plot.

col.sig.threshold

one color for the significance threshold

...

not in use

Value

a plot for viewing vGWAS result.

References

Shen, X., Pettersson, M., Ronnegard, L. and Carlborg, O. (2011): Inheritance beyond plain heritability: variance-controlling genes in Arabidopsis thaliana. PLoS Genetics, 8, e1002839.

See also

Author

Xia Shen

Examples

# \donttest{
# ----- load data ----- #
data(pheno)
data(geno)
data(chr)
data(map)
# ----- variance GWA scan ----- #
vgwa <- vGWAS(phenotype = pheno, geno.matrix = geno,
marker.map = map, chr.index = chr, pb = FALSE)
#> Error in vGWAS(phenotype = pheno, geno.matrix = geno, marker.map = map,     chr.index = chr, pb = FALSE): unused argument (pb = FALSE)
# ----- visualize the scan ----- #
plot(vgwa)
#> Error in eval(expr, envir, enclos): object 'vgwa' not found
summary(vgwa)
#> Error in eval(expr, envir, enclos): object 'vgwa' not found
# ----- calculate the variance explained by the strongest marker ----- #
vGWAS.variance(phenotype = pheno,
marker.genotype = geno[,vgwa$p.value == min(vgwa$p.value)])
#> Error in eval(expr, envir, enclos): object 'vgwa' not found
# ----- genomic control ----- #
vgwa2 <- vGWAS.gc(vgwa)
#> Error in eval(expr, envir, enclos): object 'vgwa' not found
plot(vgwa2)
#> Error in eval(expr, envir, enclos): object 'vgwa2' not found
summary(vgwa2)
#> Error in eval(expr, envir, enclos): object 'vgwa2' not found
# }