Variance Genome-wide Association

Variance Genome-wide association for using nonparametric variance test

vGWAS(phenotype, geno.matrix, kruskal.test = FALSE,
marker.map = NULL, chr.index = NULL, pB = TRUE)

Arguments

phenotype

a numeric or logical vector of the phenotyic values. See Examples.

geno.matrix

a matrix or data.frame with individuals as rows and markers as columns. The marker genotypes for each marker are coded as one column. See Examples.

kruskal.test

a logical value specifying whether to use Kruskal-Wallis statistic. The default option is FALSE, i.e., the usual ANOVA statistic is used in place of Kruskal-Wallis statistic.

marker.map

a numeric vector giving the marker map positions for each chromosome. See Examples.

chr.index

a numeric vector giving the chromosome index for each marker. See Examples.

pB

show progress bar

Value

a data.frame containing columns of marker names, chromosome indices, marker.map positions, test statistic values, and p.value for each position.

References

Shen, X., Pettersson, M., Ronnegard, L. and Carlborg, O. (2011): Inheritance beyond plain heritability: variance-controlling genes in Arabidopsis thaliana. PLoS Genetics, 8, e1002839.

Ronnegard, L., Shen, X. and Alam, M. (2010): hglm: A Package for Fitting Hierarchical Generalized Linear Models. The R Journal, 2(2), 20-28.

See also

Author

Xia Shen

Examples

# \donttest{
# ----- load data ----- #
data(pheno)
data(geno)
data(chr)
data(map)
# ----- variance GWA scan ----- #
vgwa <- vGWAS(phenotype = pheno, geno.matrix = geno,
marker.map = map, chr.index = chr, pb = FALSE)
#> Error in vGWAS(phenotype = pheno, geno.matrix = geno, marker.map = map,     chr.index = chr, pb = FALSE): unused argument (pb = FALSE)
# ----- visualize the scan ----- #
plot(vgwa)
#> Error in eval(expr, envir, enclos): object 'vgwa' not found
summary(vgwa)
#> Error in eval(expr, envir, enclos): object 'vgwa' not found
# ----- calculate the variance explained by the strongest marker ----- #
vGWAS.variance(phenotype = pheno,
marker.genotype = geno[,vgwa$p.value == min(vgwa$p.value)])
#> Error in eval(expr, envir, enclos): object 'vgwa' not found
# ----- genomic control ----- #
vgwa2 <- vGWAS.gc(vgwa)
#> Error in eval(expr, envir, enclos): object 'vgwa' not found
plot(vgwa2)
#> Error in eval(expr, envir, enclos): object 'vgwa2' not found
summary(vgwa2)
#> Error in eval(expr, envir, enclos): object 'vgwa2' not found
# }