Brown-Forsythe's Test of Equality of Variances

The function performs the robust Brown-Forsythe test using the group medians. Instead of the ANOVA statistic, the Kruskal-Wallis ANOVA may also be applied using this function.

brown.forsythe.test(y, group, kruskal.test=FALSE)

Arguments

y
a numeric vector of data values.

group
factor of the data.

kruskal.test
a logical value specifying whether to use Kruskal-Wallis statistic. The default option is FALSE, i.e., the usual ANOVA statistic is used in place of Kruskal-Wallis statistic.

Value

A list with the following numeric components.

statistic

the value of the test statistic.

p.value

the p-value of the test.

method

type of test performed.

data.name

a character string giving the name of the data.

Details

Levene (1960) proposed a test for homogeneity of variances in k groups which is based on the ANOVA statistic applied to absolute deviations of observations from the corresponding group mean. The robust Brown-Forsythe version of the Levene-type test substites the group mean by the group median in the classical Levene statistic.

Note

Modified from the lawstat package.

References

Brown, M. B. and Forsythe, A.B. (1974). Robust tests for equality of variances. Journal of the American Statistical Association, 69, 364-367.

Levene, H. (1960). Robust Tests for Equality of Variances, in Contributions to Probability and Statistics, ed. I. Olkin, Palo Alto, CA: Stanford Univ. Press.

Examples

data(pheno)
data(geno)
brown.forsythe.test(pheno, geno[,911])

#> 
#> 	Brown-Forsythe test based on the absolute deviations from the median
#> 
#> data:  pheno
#> Test Statistic = 0.452, p-value = 0.5031
#>